requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==4.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.1
django-celery-beat==2.1.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
Markdown==3.4.1
marshmallow==3.17.1
mysqlclient==2.1.1
packaging==21.3
prompt-toolkit==3.0.31
pyactiveresource==2.2.2
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ShopifyAPI==12.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
